core: Determine need to flush tlbs on paging changes via paging flag
Rather than comparing the pg_structs pointer passed to paging_create/
paging_destroy against hv_paging_structs, add a flag to struct
paging_structures hat determines if those are used for the hypervisor or
some cell. This will be useful when introducing per-cpu page tables.

This also allows to use split_hugepage for hypervisor mappings which
will be required later on. So far, the code assumed to only be called
for cell mappings.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>