x86: Hide private per-cpu section during hypervisor operation
Turn off access to the private per-cpu section before the final
activation of the hypervisor. This prevents that one CPU can potentially
speculate about the private content of per-cpu sections of others.

Stack migration has to happen earlier now, before disabling the common
access rather. So far we used the migrated stack pointer only on the
first VM exit.

Before we switch the hypervisor off, access needs to be granted again
because the page table switch and the stack migration aren't atomic.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>