core: mmio: Plug potentially dangerous race between lookup and modification
Houston, I think we had a problem here:

 CPU 1                                   CPU 2
 -----                                   -----
 copy_region(..., 1, 2)
                                         mmio_handle_access()
                                             find_region(...) = 1
 copy_region(..., 0, 1)
     mmio_handlers[1].handler = handler0
                                             use mmio_handlers[1]:
                                                 handler0(arg1)
     mmio_handlers[1].arg = arg0

And handler0 will interpret arg1 as if it were arg0, possibly using a
completely different type. Yeah, lockless is easy ... to get wrong.

What we need here to avoid taking a lock in mmio_handle_access and while
still being safe is some revision counting with a retry in case a
mmio access dispatch races with a modification of the list (could be
triggered by the guest via a reconfiguration of ivshmem mmio regions).
And this is what this patch does:

The mmio_generation counter is incremented to an odd value before
beginning a change to the list. Then the modifications are done, the
list size is adjusted, and finally the generation counter is incremented
again - all properly serialized via memory barriers. That's the writer
side.

On the reader side, we capture the generation counter before starting a
lookup. If it is already odd at that time, we wait for the modification
to finish. That is because we may otherwise capture an inconsistent
state before the writing side has incremented the counter again.

While walking the list in find_regions, we check the counter after each
iteration to avoid that we miss a shrink or an expansion. After we find
a matching entry, we capture all relevant data before we check the
generation counter once again. That ensures the the caller of
find_region will get a consistent snapshot of the matching region index,
its base address and its handler function with argument.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>